refactoring: folder structure
