+++ src/main/java/com/example/shelfofshame/book/Book.java
@@ -1,7 +1,7 @@
-package com.example.shelfofshame.entities;
+package com.example.shelfofshame.book;
 
+import com.example.shelfofshame.author.Author;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -29,10 +29,6 @@ public class Book {
     private Set<Author> authors;
     @Column
     private String description;
-    @ManyToOne
-    @JoinColumn(name = "categoryId")
-    @NotNull
-    private Category category;
     @Column
     private String coverUrl;
     @Column
