--- src/main/java/com/example/shelfofshame/entities/Category.java
@@ -1,17 +1,0 @@
-package com.example.shelfofshame.entities;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data @Builder
-@Entity
-@Table(name = "categories")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false, unique = true)
-    private String name;
-}
