+++ src/main/java/com/example/shelfofshame/user/UserMapper.java
@@ -1,8 +1,7 @@
-package com.example.shelfofshame.mappers;
+package com.example.shelfofshame.user;
 
-import com.example.shelfofshame.dto.SignupDto;
-import com.example.shelfofshame.dto.UserDto;
-import com.example.shelfofshame.entities.User;
+import com.example.shelfofshame.user.dto.SignupDto;
+import com.example.shelfofshame.user.dto.UserDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
