+++ src/main/java/com/example/shelfofshame/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.example.shelfofshame.dto;
+package com.example.shelfofshame.user.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -12,11 +12,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 @Data
-public class SignupDto {
+public class UserDto {
+    private Long id;
     @Email(message="Enter a valid email") @NotBlank(message="Email is required")
     private String email;
     @Size(min = 3, max = 20, message="Username must be between 3 and 20 characters")
     private String username;
-    @Size(min=8, max=128, message="Password must be between 8 and 128 characters")
-    private String password;
+    private String token;
 }
